Util: Make UString not taking char const*
To make sure that strlen is not implicitly called at runtime
(potentially overflowing a string and relying on the fact that it is
null-terminated), UString constructors now take only string_view or
array, so that the size is already known when calling the constructor.

See https://github.com/SerenityOS/serenity/pull/14555 for more details
why const char* is bad here.